Added static_cast's to avoid compiler warnings about conversions between 32-bit and 64-bit ints

+++ LvqEmn/LvqLib/AbstractLvqModel.h
@@ -39,7 +39,7 @@ public:
 	PMatrix const & projectionMatrix() const {return P;}
 	double projectionNorm() const { return projectionSquareNorm(P);  }
 	void normalizeProjection() { normalizeMatrix(P); }
-	virtual int Dimensions() const {return P.cols();}
+	virtual int Dimensions() const {return static_cast<int>(P.cols());}
 
 	virtual MatrixXd GetProjectedPrototypes() const=0;
 	virtual std::vector<int> GetPrototypeLabels() const=0;
