+++ LvqEmn/LvqLib/LvqDataset.h
@@ -22,5 +22,5 @@ public:
 	double ErrorRate(AbstractLvqModel const * model) const;
 	PMatrix ProjectPoints(AbstractProjectionLvqModel const * model) const;
 	size_t MemAllocEstimate() const;
-	int dimensions() const { return trainPoints.rows();}
+	int dimensions() const { return static_cast<int>(trainPoints.rows());}
 };
